refactor: turn directive into component
