+++ src/component.js
@@ -1,0 +1,12 @@
+import controller from './controller';
+import template from './template.html';
+
+export default {
+  bindings: {
+    ngModel: '=?',
+    userService: '=?',
+    mondialRelayService: '=?',
+  },
+  template,
+  controller,
+};
